pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "aims_PAX"
version = "0.1"
description = "Active learning for FHI-aims"
readme = "README.md"
requires-python = ">=3.9"
license = { text = "Placeholder" }
authors = [
    { name = "Tobias Henkes", email = "tobias.henkes@uni.lu" }
]
keywords = []
classifiers = [
    "Programming Language :: Python :: 3"
]
dependencies = [
    "ase==3.23.0",
    "numpy==1.26.4",
    "mace-torch==0.3.13",
    "asi4py==1.3.18",
    "torch==2.7.1",
    "torchvision==0.22.1",
    "torchaudio==2.7.1",
    "pyyaml==6.0.1",
]

[project.optional-dependencies]
parsl = ["parsl==2024.12.16"]

[project.urls]
homepage = "https://github.com/tohenkes/aims_PAX"

[tool.setuptools.packages.find]
where = ["src"]

[project.scripts]
aims_PAX = "aims_PAX.cli.__main__:main"
aims_PAX-initial-ds = "aims_PAX.cli.create_initial_ds:main"
aims_PAX-al = "aims_PAX.cli.al_procedure_only:main"
